CB-22369 CcmV2ManagementClientTest is flaky
Spring Retry Template got exhausted in our PR's integrated test part. It could not be
reproduced though, it is remedied by refactoring lambda calls of CCMv2 client to
a separate component's method calls.
Testing is splitted into 2 parts: CCMv2ClientTest contains the "happy" paths of the retrying
client, while the CCMv2RetryingClientTest contains the "retry exhausted" paths.
